Made Monte_Carlo_Simulation class portable

+++ elec631/20171116_0107_duck_shooting/include/Monte_Carlo_Simulation.hpp
@@ -1,27 +1,27 @@
 #ifndef MONTE_CARLO_SIMULATION_H_20171126_1728
 #define MONTE_CARLO_SIMULATION_H_20171126_1728
-
+#include <iostream>
 namespace Monte_Carlo_Simulation_NS{
     class Monte_Carlo_Simulation{
         private:
-            int noTests_;
-            int noHunters_;
-            int spacing_;
+            uint64_t noTests_;
+            uint64_t noHunters_;
+            uint64_t spacing_;
 
         public:
             /** Default constructor */
             Monte_Carlo_Simulation();
 
             /** Three Argument Constructor*/
-            Monte_Carlo_Simulation(const int, const int, const int);
+            Monte_Carlo_Simulation(const uint64_t, const uint64_t, const uint64_t);
 
             /** Default destructor */
             ~Monte_Carlo_Simulation();
 
             /** Getters */
-            const int NoTests_() const;
-            const int NoHunters_() const;
-            const int Spacing_() const;
+            const uint64_t NoTests_() const;
+            const uint64_t NoHunters_() const;
+            const uint64_t Spacing_() const;
 
             /** Simulation */
             void simulate();
